refactor: extract network intake builder
